energy bar has smoother transition now
+the bar now intellegently glides between states
+some related refactoring
?attempted to give the same effect to the charge bar, but failed repeatedly...

+++ ponies/abilities/EnergyUsePiece.java
@@ -5,9 +5,9 @@ public class EnergyUsePiece<T> {
 	
 	public int id;
 	public T cost;
-	public int timestamp;
+	public long timestamp;
 	
-	EnergyUsePiece(int id, T cost, int timestamp)
+	EnergyUsePiece(int id, T cost, long timestamp)
 	{
 		this.id = id;
 		this.cost = cost;
